Dropdown Fix
Dropdown now overlaps instead of pushing elements down. Also slightly changed implementation of typings.
