+++ src/components/shared/booleanDropdown.tsx
@@ -3,12 +3,17 @@ import '../../styles/booleanDropdown.scss';
 
 const options = ['True', 'False'];
 
+export enum Position {
+  Top = 'top',
+  Middle = 'middle',
+  Bottom = 'bottom'
+}
+
 interface BooleanDropdownProps {
   change: () => void;
+  position: Position;
 }
 
-function BooleanDropdown(props:BooleanDropdownProps):JSX.Element  {
-  return <Dropdown className='drop' options={options} value='False' arrowClassName='arrow' onChange={props.change}/>;
+export function BooleanDropdown(props:BooleanDropdownProps):JSX.Element  {
+  return <Dropdown className={'drop '+props.position} options={options} value='False' arrowClassName='arrow' onChange={props.change}/>;
 }
-
-export default BooleanDropdown;
