--- src/components/shared/typings.ts
@@ -1,10 +1,0 @@
-export enum Color {
-  White = 'white',
-  Orange = 'orange',
-  Blue = 'blue',
-}
-
-interface bitOfCode {
-  [key: string]: Color;
-}
-export type lineOfCode = bitOfCode[];
