+++ src/styles/python.scss
@@ -35,6 +35,12 @@
   width: 180px;
 }
 
+#dropdowns {
+  height: 90px;
+  position: absolute;
+  width: 70px;
+}
+
 #problem {
   display: flex;
   margin-top: 40px;
